+++ plugins/interfaces/qfrdrimagemask.h
@@ -22,14 +22,14 @@ class QFRDRImageMaskInterface {
         virtual QString maskToString() const=0;
         virtual void maskClear()=0;
         virtual void maskSetAll()=0;
-        virtual void maskUnset(uint16_t x, uint16_t y)=0;
-        virtual void maskSet(uint16_t x, uint16_t y, bool value=true)=0;
-        virtual void maskToggle(uint16_t x, uint16_t y)=0;
+        virtual void maskUnset(uint32_t x, uint32_t y)=0;
+        virtual void maskSet(uint32_t x, uint32_t y, bool value=true)=0;
+        virtual void maskToggle(uint32_t x, uint32_t y)=0;
         virtual void maskInvert()=0;
-        virtual bool maskGet(uint16_t x, uint16_t y) const =0;
+        virtual bool maskGet(uint32_t x, uint32_t y) const =0;
         virtual bool* maskGet() const=0;
-        virtual uint16_t maskGetWidth() const=0;
-        virtual uint16_t maskGetHeight() const=0;
+        virtual uint32_t maskGetWidth() const=0;
+        virtual uint32_t maskGetHeight() const=0;
         virtual void maskMaskChangedEvent()=0;
 
 };
